Add filtering ability to the list command
This lets the user specify a list of field-value pairs at the end
of the command which will be used to filter the results.

This feature will also be added to the delete and add commands because
the are some api objects that don't have uuids, therefore this will be
the only way to select them to perform an action on them.

Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>