CLI: fix autocompletion bugs
The new complete_with_object() method in FieldAssignmentBase was calling
complete() using an object at the wrong depth level in the object hierarchy.
In the example, this meant calling complete on the Router object instead
of the RouterPort object.

Also, list filters didn't have an implemented autocomplete, because they
were using the redundant and incompete Seq class. This patch removes that
class and uses Concat in its place, which implements complete() correctly.

Fixes MN-189
Related to MN-192

Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>