CLI: fix bug in contructor parameter ordering
A new 'unsetter' parameter was introduced in SingleAttr, however some
subclasses that don't need it didn't add it to their call to the superclass.
The were sensitive to this change because they call the super class passing
arguments by ordering, not by name.

Fixed this by moving 'unsetter' to the last place in the list, which makes
sense because it's the least likely argument to be passed: only device
ports have a special-purposed 'unsetter'.

Fixes MN-192.

Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>