CLI exits on first command when a credential is invalid.
Currently CLI goes back to a normal command line loop showing a
prompt even when a previous command failed because of an invalid
credential. This is confusing as it can appear as if authentication
was successful and CLI was ready to take commands. The new behavior
is that CLI will first run a test command to see if a credential is
valid, and exit right after if invalid, displaying a corresponding
error message.

Fix MN-718

Change-Id: I2ceb940f35c7418edee3601a7fc55874e24436dd