Add the setuptools package configuration
Currently we can only use the debian package or the RPM package to
install python-midonetclient. It's not a trivial process to create these
packages and that makes adding new features harder. And it's not easy to
try this package out on Mac OSX.

This patch addds the setuptools package configuration, which makes us to
install python-midonetclient easier. We can also use virtualenv/venv to
separate the installation environments out from each other.

Because python-midonetclient is a public project, we can register this module
on PyPI. Feel free to register with the following command:

    $ python setup.py register

Signed-off-by: Taku Fukushima <tfukushima@midokura.com>