VTEP binding delete accepts only a single binding.
This fixes the issue with the previous VTEP binding deletion patch
that it allowed the user to delete multipe VTEP bindings at once,
which is not consistent with the existing deletion semantics. This
patch enforces the uniqueness of the binding deletion by reusing the
ObjectFromCollection construct. It also extends the accepted deletion
expressions to include "delete PARENT-COLLECTION CHILD-COLLECTION".

Fix MN-1824