Warns a user when specifying a VTEP name.
This patch implements 'writeable' field of SingleAttr. If the field is
set to False, SingleAttr throws a newly defined UserException when a
user attempts to set a value to a read-only attribute such as VTEP
name and description, causing the command execution to abort and to go
back to the command line evaluation loop.

Please note that though this 'writeable' field hasn't been
implemented, many entities are specifying various fields 'writeable',
thus implementing this feature may have unforeseen impacts.

Fix MN-2510